+++ Anime/app/components/Tab.js
@@ -15,7 +15,7 @@ const styles = StyleSheet.create({
     flex: 1,
     alignItems: "center",
     justifyContent: "center",
-    padding: 13,
+    padding: 10,
   },
 });