+++ Anime/app/components/TabBar.js
@@ -1,23 +1,49 @@
-import { StyleSheet, View, Dimensions, StatusBar } from "react-native";
-import React, { useState } from "react";
+import {
+  StyleSheet,
+  View,
+  Dimensions,
+  StatusBar,
+  BackHandler,
+  Alert,
+} from "react-native";
+import React, { useState, useEffect } from "react";
 import Tab from "./Tab";
+import { useBackHandler } from "@react-native-community/hooks";
 
 const { width } = Dimensions.get("screen");
-
 const TabBar = ({ state, navigation }) => {
+  //Handles the back press Action with alert box
+  const backActionHandler = () => {
+    Alert.alert("Confirm", "are you sure you want to exit app ?", [
+      {
+        text: "Cancel",
+        onPress: () => null,
+        style: "destructive",
+      },
+      {
+        text: "Yes",
+        onPress: () => BackHandler.exitApp(),
+      },
+    ]);
+    return true;
+  };
   const [selected, setSelected] = useState("Home");
   const { routes } = state;
+
+  //Helps to render the color dynamically
   const renderColor = (currentTab) =>
     currentTab === selected ? "#2fceed" : "#000";
 
   const handlePress = (activeTab, index) => {
-    console.log(activeTab);
     if (state.index !== index) {
       setSelected(activeTab);
       navigation.navigate(activeTab);
       activeTab === index;
     }
   };
+
+  //custom function from react-native community to handle back press
+  useBackHandler(backActionHandler);
 
   return (
     <View style={styles.wrapper}>
@@ -40,7 +66,7 @@ const TabBar = ({ state, navigation }) => {
 const styles = StyleSheet.create({
   wrapper: {
     position: "absolute",
-    bottom: 20,
+    bottom: 5,
     width,
     alignItems: "center",
     justifyContent: "center",
@@ -49,10 +75,13 @@ const styles = StyleSheet.create({
     flexDirection: "row",
     justifyContent: "space-between",
     backgroundColor: "#fff",
-    width: 250,
+    width: 350,
     borderRadius: 100,
-    elevation: 5,
-    shadowColor: "#18606e",
+    elevation: 10,
+    shadowColor: "#000",
+  },
+  cancel: {
+    color: "red",
   },
 });
 export default TabBar;