+++ Anime/CHANGELOG.md
@@ -1,4 +1,8 @@
 ### V1.0.0
 
 - Created an app
-- Created bottom navigation bar
+- Created bottom navigation bar with three screens
+- Added Web View for home page
+- Added back to previous page on back press to webview
+- Modified bottom navigation bar
+- Added alert on exit